+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- Account決済ルーターをPaymentCard決済ルーターに統合
+
 ### Deprecated
 
 ### Removed